Feat: Integração e tela de loguin e cadastro

+++ backend/app/routers/game_router.py
@@ -1,5 +1,6 @@
-from fastapi import APIRouter, HTTPException
+from fastapi import APIRouter, HTTPException, Depends
 from .. import database
+from .auth_router import verify_token
 
 router = APIRouter(
     prefix="/games",
@@ -9,13 +10,19 @@ router = APIRouter(
 db = database.db
 
 @router.get("/{user_id}")
-def list_games(user_id: str):
-
+def list_games(user_id: str, token: dict = Depends(verify_token)):
+    
+    if token['uid'] != user_id:
+        raise HTTPException(status_code=403, detail="Acesso negado.")
+        
     docs = db.collection("users").document(user_id).collection("games").stream()
     return [doc.to_dict() for doc in docs]
 
 @router.post("/{user_id}")
-def add_game(user_id: str, game: dict):
+def add_game(user_id: str, game: dict, token: dict = Depends(verify_token)): # <--- Faltava isso aqui
+
+    if token['uid'] != user_id:
+        raise HTTPException(status_code=403, detail="Acesso negado.")
 
     try:
         ref = db.collection("users").document(user_id).collection("games").document()
@@ -25,7 +32,10 @@ def add_game(user_id: str, game: dict):
         raise HTTPException(status_code=500, detail=str(e))
 
 @router.put("/{user_id}/{game_id}")
-def update_game(user_id: str, game_id: str, game: dict):
+def update_game(user_id: str, game_id: str, game: dict, token: dict = Depends(verify_token)): # <--- E aqui
+
+    if token['uid'] != user_id:
+        raise HTTPException(status_code=403, detail="Acesso negado.")
 
     try:
         ref = db.collection("users").document(user_id).collection("games").document(game_id)
@@ -35,7 +45,10 @@ def update_game(user_id: str, game_id: str, game: dict):
         raise HTTPException(status_code=500, detail=str(e))
 
 @router.delete("/{user_id}/{game_id}")
-def delete_game(user_id: str, game_id: str):
+def delete_game(user_id: str, game_id: str, token: dict = Depends(verify_token)): # <--- E aqui também
+
+    if token['uid'] != user_id:
+        raise HTTPException(status_code=403, detail="Acesso negado.")
 
     try:
         db.collection("users").document(user_id).collection("games").document(game_id).delete()
